Add Unsigned comparisons to JitBuilder
Unsigned compares are really useful, but currently not provided
by JitBuilder in IlBuilder or BytecodeBuilder. The OMR compiler
supports unsigned comparisons, so there's no reason the shouldn't
be availble to JitBuilder.

I added Unsigned varieties for all comparisons in IlBuilder and
BytecodeBuilder. In the process, I managed to simplify some of
the logic used to generate conditional operations by consolidating
on TR_ComparisonType as the mechanism to get the appropriate
comparison opcode based on the datatype and whether the compare
should be signed/unsigned.

Issue: #833
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>