Add conditionals code sample for JitBuilder
This commit adds a new code sample demonstrating how IfAnd and
IfOr can be used. Under the covers, many of the signed and unsigned
comparison services are used as well. The compiled method simply
performance a range test using four different equivalent approaches
and verifies that the same result is computed by all four approaches.
The test then calls that method with a variety of positive and
negative numbers.

Issue: #833
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>