Add NOPGuard service to IlBuilder
Takes two arguments: the first is a builder object representing
the code that will run as fall-through from the guard, and the
second is the builder object representing the code that should
execute if the assumption is later invalidated. NOPGuard()
returns a JitAssumption (currently a uint32_t) value that can
be later used to invalidate the guard.

Both the guarded and failed path will, by default, merge after
executing. Services called after NOPGuard on the same builder
will execute after this merge point.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>