Add VectorLoopBuilder class to JitBuilder
Provide a convenience class for constructing vectorized loops
which automatically manages both the vector loop and its
corresponding residue loop. Services that can be vectorized
are mapped into both loop bodies and related in a map so that
only one value needs to be used by the client.

Add a VectorForLoop() service to IlBuilder that automatically
creates the entire vector loop and residue loop control
structure and returns a new VectorLoopBuilder to be used to
inject operations into this loop structure.

Also enable outer loop vectorization by introducing a ForLoop
service in VectorLoopBuilder that creates a new VectorLoopBuilder
object to represent the two "scalar" inner loops (one in the
outer vectorized loop, and one in the outer residue loop). To
properly support inner for loops, VectorLoopBuilder objects
maintain their parent chain so that values can be properly
mapped between scalar and vector loops correctly, no matter
in which loop of the loop nest the value is created.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>