Add Vector Matrix Multiply code sample for JitBuilder
This sample shows how to use the VectorForLoop service to
automatically generate a fully vectorized loop along with
residue loop, only requiring the client to specify the
loop body code once (it is automatically written into the
residue loop at the same time it populates the vector
loop).

One of the coolest parts is that this service works even
if there are inner loops: this sample shows how to
vectorize the middle loop in the standard ijk form of
matrix multiplication.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>