Use correct node for value in VectorStore
VectorStore properly looks to see if the value being
provided is a scalar and therefore nees to be splatted
to a vector, but then the actual store node generated
uses the original value (whoops!). Simple fix is to
use the (possibly splatted) updated node value.

Also added a missing TraceIL call that will help make it
easier to track what's going on in the traceIlGen output.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>