Simplify compiler basic memory allocation classes (first stage)
The lowest level memory allocation classes in the OMR compiler
have a complicated hierarchy and relationship that many people
have struggled to understand and maintain. The complexity came
about trying to hold a unified picture of memory allocation
between OMR (which just uses a raw malloc allocator) and OpenJ9
(which uses J9MemorySegments allocated from the VM a.k.a.
"the system"). Unfortunately, this unity did not materialize
and so these lower level allocation classes are not actually
shared with OpenJ9.

The set of changes in this commit is motivated by the effort
to try to compile the basic JitBuilder files in compiler/ilgen
as part of OpenJ9. In this commit, I brought over the code from
OpenJ9's SystemSegmentProvider and refactored it so that it
could operate in OMR using TR::MemorySegments rather than
J9MemorySegments. I changed its name to SegmentAllocator and
declare the SegmentAllocator as the base layer for the higher
level allocation support for e.g. regions and STL and other
objects in the compiler component.

This commit is the first stage in a series of coordinated
changes to OMR and OpenJ9 to migrate from the "old" memory
allocation scheme to the "new" memor allocation scheme. In
this change, *most* of the new code is implemented inside
a new compiler/env/mem directory which contains any files
from compiler/env that required substantial modification.
This commit also introduces a temporary macro called,
uninspiringly, OLD_MEMORY: if OLD_MEMORY is defined then
the old OMR memory allocation classes will be built rather
than the new ones. By default after this commit, OMR will
build with the *new* memory allocation classes (e.g. into
jitbuilder and compilertest). Most consumers of the OMR
compiler should experience *zero* functional differences
as the memory classes being changed are at the lowest levels
of the compiler implementation (well below than the facilities
most compiler developers use directly).

The SegmentAllocator will hold a BackingMemoryAllocator which
will itself hold a RawAllocator.  In fact, the
BackingMemoryAllocator in OMR is an only slightly beefed up
RawAllocator which returns memory wrapped in a TR::MemorySegment.
The SegmentPool code was removed because nobody was using it.
he DebugSegmentProvider code has been moved into DebugRawAllocator
(subclass of TR::RawAllocator) as well as DebugSegmentAllocator
(subclass of TR::SegmentAllocator).

I added some tracing output to the SegmentAllocator and
BackingMemoryAllocator to track segments as they are allocated
and deallocated (see MEMLOG macros in each file). This debug
output uses printfs due to the primordial environment of the
memory allocators, and is activated by environment variables
OMRDebug_BackingMemory=1 or 2 or 3    (increasing detail)
OMRDebug_SegmentAllocator=1 or 2 or 3 (increasing detail)

While debugging the memory allocation layers, I discovered that
the original list management for segments was less efficient than
it could be, in that new segments were typically added at the
end of lists but searched starting from the beginning. This commit
changes the order of search to require fewer probes (less than
half the probes required in some very unscientific studies with
very simple compilations). I doubt it will have any performance
impact in most workloads but it felt better to fix it.

BackingMemoryAllocator, SegmentAllocator, and DebugSegmentAllocator
are all extensible classes to facilitate a similar refactoring
extravaganza to be inflicted upon OpenJ9 very soon.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>