Update compiler memory allocation documentation
The compiler's lowest level memory allocation implementation
is changing so provide documention for the new scheme. But since
the old implementation is still available via the OLD_MEMORY macro,
the older documentation is not removed, but moved to a file prefixed
by "Old" and with comments at the top directing towards the new
file. Once all downstream projects have stopped using the old scheme,
this older documentation will be removed.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>