Create new compiler SegmentAllocator interface
SegmentAllocator is an abstract class that will be used by the rest of
the compiler to interact with MemorySegment allocators (e.g. concrete
subclasses like NaiveSegmentAllocator, NativeDebugSegmentAllocator,
SegmentCache, etc.). SegmentAllocator objects must track all memory they
allocate and ensure that any memory that has not been explicitly
deallocated is deallocated when the SegmentAllocator object is destructed.

SegmentAllocator objects are not required to be thread-safe so explicit
locking is needed if multiple threads are able to access a single
SegmentAllocator object simultaneously.

The main distinguishing feature of the SegmentAllocator interface versus
RawAllocator is that SegmentAllocator deals in TR::MemorySegment objects
rather than RawMemory pointers.

SegmentAllocator is part of the new memory allocation implementation which,
until refactoring is complete, will live in the compiler/env/newmemory
directory.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>