Create new compiler NaiveSegmentAllocator class
NaiveSegmentAllocator<RawAllocator> is a SegmentAllocator that straight-forwardly
allocates MemorySegments when allocate() is called and frees them when deallocate()
is called. The RawAllocator template paramater is used to embed an object of
a particular TR::RawAllocator object, so this must name a concrete class
implementing the TR::RawAllocator interface.

The OMR implementation of NaiveSegmentAllocator uses its RawAllocator to allocate
both the underlying memory as well as the TR::MemorySegment object. This class
is designed, however, to be subclassable to override the protected member
functions directAllocate() and directDeallocate() to provide code to
allocate/deallocate memory without necessarily using the embedded RawAllocator
directly.

This allocator rounds all allocations up to a "minimum" allocation size to avoid
fragmentation from the backing RawAllocator. In typical configurations, this minimum
allocation size will be set to something like 64KB or 1MB. An allocation limit can be
optionally set to prevent excessive memory allocation.

Every call to allocate() will definitely allocate a MemorySegment and calls to
deallocate will definitely deallocate both the memory and the TR::MemorySegment
object. Projects may subclass OMR::NaiveSegmentAllocator to change this behaviour,
and even other subclasses in OMR (e.g. OMR::NaiveDebugSegmentAllocator) may not
always deallocate segments when deallocate() is called.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>