Create new compiler NaiveDebugSegmentAllocator class
The NaiveDebugSegmentAllocator class provides a facility for verifying the
correctness of compiler scratch memory use.

Using the native facilities available on each platform, a
NaiveDebugSegmentAllocator provides an alternative allocation mechanism
for the compiler's scratch memory regions. Instead of releasing virtual
address segments back to the operating system, this implementation instead
either remaps the segment (freeing the underlying physical pages and
changing the memory protection to trap on any access [preferred]), or,
if such facilities are not available, paints the memory segment with a
value that should cause pointer dereferences to be unaligned, and resolve
to the high-half of the virtual address space often reserved for kernel /
supervisor use.

Once the compiler has been modified to enable stateful RawAllocator objects,
the functionality in this class is expected to sink completely into
DebugAllocator and this class will no longer be required on top of
OMR::NaiveSegmentAllocator<OMR::DebugAllocator>. Until then, it's needed
to track the "deallocated" segments until destruction time.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>