Create new compiler SegmentCache class
A SegmentCache is a TR::SegmentAllocator that attempts to reuse
allocated segments as much as possible. The SegmentCache will only
attempt to reuse segments below a certain size (cacheableSizeThreshold)
and will cut up deallocated segments into smaller segments of
cacheBlockSize. Allocations smaller than cacheBlockSize will be rounded
up to the cacheBlockSize to prevent fragmentation. Any allocation
request above the cacheableThresholdSize will be directly allocated
sing the provided SegmentAllocator (which is typically a
NativeSegmentAllocator but does not have to be). Such large segments
will also be directly deallocated rather than being cut up to be
placed on the free segment list.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>