Create new compiler CachingSegmentAllocator class
A CachingSegmentAllocator is a TR::SegmentAllocator that attempts
to avoid allocating segments by first using any segments available
from a SegmentCache. Only if the SegmentCache does not provide a
segment will the CachingSegmentAllocator attempt to allocate using
a downstream TR::SegmentAllocator object. The use of this class
enables a SegmentCache to be used to allocate long-lived objects
(via its own SegmentAllocator) as well as a shorter-lived
SegmentAllocator to be used when the SegmentCache is full.

The typical use case is to use a long-lived SegmentCache to hold a
small number of large segments for reuse across all compilations
for a period of time. "Long-lived" here means lives across
compilations. A second SegmentCache can then be created per
compilation. This shorter-lived SegmentCache needs to allocate
segments somehow: it can use a CachingSegmentAllocator to prefer
using segments from the long- lived cache until they are used up. Any
subsequent segments will be allocated from the CachingSegmentAllocator's
downstream segment allocator (which can have compilation lifetime
as well). Eclipse OpenJ9 uses this scheme, but there's nothing
OpenJ9-specific about this approach..

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>