Update MemorySegment for new memory allocation classes
MemorySegment needs a "rewind()" capability under the new memory
allocation classes for cases where the raw memory has been
successfully allocated from a segment but either the TR::MemorySegment
object itself cannot be allocated or the TR::MemorySegment
object was not successfully recorded into a std::deque. In
these cases, the earlier memory allocation should be undone.

Rewind is a dangerous mechanism: it assumes the last allocation needs
to be "rewound" so it just decrements the _allocated index in the
same way it increments it to allocate space from the segment. This
function is dangerous if not used carefully, so it is commented
profusely to hopefully avoid misuse.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>