Add C and Java apigen cmake support
In-process updates to cmake support to be able to build the C
and Java API

CPP bindings are built automatically. Can be deactivated by adding
-DOMR_JITBUILDER_BINDING_CPP=OFF to the cmake command.

C bindings are only built if -DOMR_JITBUILDER_BINDING_C=ON is added
to cmake command.

Java bindings are only built if -DOMR_JITBUILDER_BINDING_JAVA=ON is
added to cmake command.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>