Enable omrtrace to compile with CMAKE_BUILD_TYPE=Release
If you try to build OMR using the cmake option -DCMAKE_BUILD_TYPE=Release
then, on some systems, you can encounter a problem trying to compile the
omrtrace component. You'll see a fairly cryptic error message like:

/usr/include/x86_64-linux-gnu/bits/string3.h:126:71: error: call to char* __builtin___strncpy_chk(char*, const char*, long unsigned int, long unsigned int) will always overflow destination buffer [-Werror]
  return __builtin___strncpy_chk (__dest, __src, __len, __bos (__dest));

There is a known fix for this problem, but it is currently commented out and
needs to account for changes that have been made since that code was written.

With this change, we should be able to reliably build OMR with build type
Release.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>