JitBuilder2 (JB2) prototype
First draft prototype of an enhancement to the JitBuilder API to
evolve it into an intermediate representation (IR) that can be
analyzed and transformed ahead of being converted to native
instructions using the OMR compiler. More documentation will
follow, but this commit includes the prototype implementation
located, for the moment, in the jitbuilder/jbil directory.
The use of the prototype is demonstrated by a slightly modified
version of the matrix multiply code sample. The demonstration
uses almost the same raw code sample but showcases some of
what can be done with the prototype. A PrettyPrinter class
can generate a readable log of the current state of the
JB2 representation. The PrettyPrinter class implements a
Visitor class that uses a worklist to traverse all the
builder objects and all the operations in a function.
Another such class is Transformer, which allows builders
and operations to be replaced as they are visited. Three
Transformer subclasses are provided: an extremely simple
AppendBuilderInliner that replaces an AppendBuilder operation
with the builder it appends if that builder is not the
target of any branching operation, a DialectReducer that
at the moment converts a ForLoop operation into lower
level conditional branches, and a CodeGenerator that
calls through the original JitBuilder API to the OMR
compiler to generate native code for a function.

Also included is a more elaborate example of how a new Type
(Complex) and new Operations (Conjugate, ConstComplex, and
Magnitude) can be added to the core JitBuilder2. These
types and operations are used in the ComplexMatMult code
sample in the same directory which is remarkably similar,
almost identical in fact, to the original code sample. To
generate native code, this new "dialect" must be reduced
into types and operations that the OMR compiler knows about.
The ComplexReducer subclass of Transformer performs this
reduction by mapping Complex values to pairs of Double
values and expanding operations with Complex operands or
results into pairs of Operations that operate on the
real and imaginary parts of each Complex value.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>