Add initial jb2 implementation
This commit is massive for two reasons: 1) it adds the jb2 (JitBuilder 2.0)
api into OMR, which is a layer of infrastructure on top of OMR, 2) since jb2
depends on jitbuilder, we actually copy the jitbuilder directory to jb2 and
then add the jb2 api into that directory. This approach was chosen so that
the original jitbuilder APIs can continue to be used by any external users
who have built on top of them, while the jitbuilder implementation that jb2
relies on can be tailored, over time, specifically for what jb2 needs without
affecting other jitbuilder users.

The new jb2 api is stored in jb2/api so that it does not overlap at all with
the jitbuilder implementation. Over time, I expect the jitbuilder implementation
to be simplified so that it does not define an "intermediate" API between jb2
and the OMR compiler (rather, the jb2 compiler should eventually directly generate
OMR compiler IL, or potentially even no longer use the OMR compiler infrastructure
(this is NOT an established goal at the moment, merely a direction we could choose
to take it in).

Documentation on jb2 is provided in jb2/doc directory.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>