Convert Operation vector to explicit linked list of Operations
Builder holds a vector of Operations, which can be a clumsy way to
store and manipulate the sequence of Operations for the Builder.
This commit adds explicit _next, _prev to Operation and adds
_firstOperation, _nextOperation, and _operationCount to Builder
for maintaining the sequence. Also add "unlink" and "replace"
to Operation which simplify other services like Transformer.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>