Remove condition on uninitialized value
On first request in a ResolvedMethodSymbol, the auto sym refs array is
alllocated based on the shape of the underlying resolved method. This
commit removes a strange condition that instead allocates space for 100
auto sym refs if "this" symbol is the compilation's jitted method
symbol. Unfortunately, this code is called while initializing the
compilation's jitted method symbol and so the field on the compilation
hasn't been initialized yet. In any case, the resolved method has
all the information needed to allocate the auto sym refs array, so
just do that all the time.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>