Jitbuilder: fix memory leaks
While running valgrind on some jb2 tests, these persistent
memory leaks were discovered in JitBuilder:
   - IlReferences are allocated but never freed
   - TypeDictionary basic Types are allocated but never freed
   - PointerTypes are allocated but never freed

This commit adds code to explicitly free the IlReferences and
basic Types when they are no longer required. It also adds
tracking of PointerTypes as they are created so they can be
collectively freed when the TypeDictionary object is destructed.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>