Update CodeGenerator API and addon classes
This commit improves how the CodeGenerator API integrates into the
extensibility features of JB2.

A new CodeGeneratorForExtension base class is defined that can be
extended by Extensions to define what the Extension needs to support
through code generation (i.e. providing type registration, support
for extension specific operations, etc.). Examples are provided
for the Core, Func, Base, and VM extensions. The jb extension
subclasses these to provide JBCodeGeneratorFor<extension>
implementations, which are then stored as addons to the
corresponding Extension object.

How are these addons managed? There is a base class called
CodeGeneratorExtensionAddon which really just records a pointer
to a CodeGeneratorForExtension object. A CodeGenerator (for example
the JBCodeGenerator in the jb extension) then provides a JB
specific subclass so that JBCodeGeneratorFor<extension> objects
can be attached and retrieved from Extension objects. For example,
JBExtension creates a JBCodeGeneratorForCore (ccg) object for
CoreExtension. This ccg object is then stored into
a JBCodeGeneratorExtensionAddon object, which is then attached to
the CoreExtension object. In this way, any JB code generator
function with access to the CoreExtension object can request
the addon for JBCodeGeneratorExtensionAddon and the returned
addon will be able to then provide a pointer to the
JBCodeGeneratorForCore object. Because CodeGeneratorForExtension
defines some virtual functions as well, it provides a way for
a CodeGenerator object like JBCodeGenerator to dispatch the handling
for Operations, Types, etc. to an extension specific object that
is also specific to JB. Multiple inheritance is avoided at some
extra cost to find and fetch the appropriate addons.

These facilities also allow an extension (like JBCodeGenerator)
to implement an API (CodeGenerator) in a way that completely
isolates the nature of "JB" from any other extension or from
the core of JB2. While it does require more classes to be defined,
these classes are mostly simple to characterize and the code
one writes to access them is fairly straight-forward once you've
done it a few times.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>