AArch64 system linkage: do not allocate frame if not needed
Track whether anything needs to be allocated into the frame and
do not bump the frame pointer if the frame size is zero.

Discovered in JitBuilder tests where what should be single instruction
functions unnecessarily bump the stack pointer by 16 and then back.
Probably not very beneficial in realistic workloads but also an easy
optimization.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>