jb2: introduce Compiler dictionaries
This commit introduces a set of dictionaries (literal, symbol, type) that are
managed at the Compiler level (so they are common to all compilations). For
example, primitive types like Int8, Address, or Float32 will be added to the
CompilerTypeDictionary by the Base extension. All Compilations will inherit
these type definitions to be managed by an as yet nonexistent set of dictionaries
that are part of the IR object. A subsequent commit will modify the existing
LiteralDictionary, SymbolDictionary, and TypeDictionary classes to become these
IR dictionaries that will be initialized from the Compiler level dictionaries.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>