jb2: prep to make all IR elements subclasses of ExtensibleIR
While writing a templated base class for the IR versions of the literal, symbol,
and type dictionaries, it became necessary to augment the IRCloner support in the
IR elements to support a virtual clone() that returns and ExtensibleIR *. But the
original type-specific clone functions (e.g. Builder *Builder::clone(...) ) are
still desirable. This commit introduces some changes to create the clone functions
into the hierarchy of IR elements that build successfully but are hacky. This will
all be cleaned up in a subsequent commit.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>