Convert KindService to return IDs and store bitvectors
KindService currently returns a 64-bit bit vector representing the
bits for a new kind along with facilities to test for membership
and equality between two kind vectors. As jb2 increases scope, we're
going to run out of bits, though. This change converts the KindService
to change Kind into a 64-bit ID (just like other structures). So
assigning a new kind simply increases the current kind ID by one,
expanding the range, in principle, to 2^64 - 1 kinds. The KindService
associates each Kind id with a BitVector (arbitrary length) that
represents all of the Kind information for that particular ID.
This change requires match() and exactMatch() to fetch the bit
vectors rather than comparing the Kind IDs.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>