Begin reintegration of jb2 into OMR project
jb2 should not be a compiler extension in an of itself: it should contain just
the jb2 apis. To this end, remove everything in the jb2 directory and recreate
jb2 as the contents of the original jb2/api directory. For now, disable all
extensions, which makes jb2 less useful, but at least the core tests can still
be run. Subsequent commits will reintegrate with either compiler or jitbuilder
to provide code generation using OMR compiler APIs via the corresponding omrgen
or jb extensions.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>