Enhance default ObjectModel.cpp implementation
jitbuilder and compilertest both have to define their own ObjectModel
classes simply to provide a definition for sizeofReferenceField()
(in both cases defined to be sizeof(char *) ).
compiler/env/OMRObjectModel.cpp already declares this function but
it's UNIMPLEMENTED forcing compiler extensions to define it for
themselves. But the base ObjectModel already provides a default
implementation for the size of a reference (32 or 64 bits depending
on the platform) so that answer should be good enough for a default
size for the reference field.

With this default function definition, we can eliminate two files
from every compiler extension.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>