Restructure optimizer for JitBuilder and compilertest
We have two competing goals for the contents of the optimizer
component: 1) we want to test everything, and 2) it would be
nice not to have to include everything if not everything is
needed. The Optimizer class currently pulls a bit of a swizzle
by using a macro to rip out all the default optimization and
group initializations so that a compiler extension can define
its own optimizer. Unfortunately, this puts all the onus onto
the compiler extension to know what it wants if it doesn't want
everything. That's what the jitbuilder compiler extension does.

Rather than this all or "tell me everything you want" approach,
we could create a couple of options available for downstream
compilers. I'm going to call them FullOptimizer (i.e. enable
everything) and SmallOptimizer (which coresponds to what jitbuilder
currently uses). In fact, FullOptimizer and SmallOptimzier are
extensible classes.  By default, Optimizer.hpp will make
TR::Optimizer derive from TR::FullOptimizer. But a compiler
extension (like jitbuilder) can now get the simpler SmallOptimizer
in one of two ways:
  1) provide an optimizer/Optimizer.hpp file that declares
       TR::Optimizer extended from TR::SmallOptimizer.
  2) (preferred) define a macro OMR_OPTIMIZER_SMALL in the build and
       exclude optimizer/FullOptimizer.cpp from the build files

By default, a compiler extension will get everything (which is
what was happening previously and some downstream projects like
OpenJ9 depend upon this behaviour). The original macro has been
removed which may introduce some downstream effects (though I'm
not explicitly aware of any projects using it).

The jitbuilder project will now follow option #2 above (defining
the macro and excluding the file). The test compiler in
fvtest/compilertest will be built with FullOptimizer so both
approaches can and will be tested within the OMR project
(which tests both jitbuider and the testcompiler by default
in its builds).

As part of this change, I also cleaned up the creation of
the Optimizer object as well as some of the handling of
the optimizationStrategy selection. OpenJ9, because it
more substantially extended the OMR optimizer, requires
some downstream changes in this area but it may be the only
downstream consumer that will encounter issues.

OpenJ9 also needs access to some of the optimization lists that
otherwise might be only referenced inside OMRFullOptimizer.cpp .
The natural thing would be to make these lists static const
(and many of them are) but some of them are needed by OpenJ9
so I have tagged these with a comment that identifies them
as /* temp? for openj9 static */ . With some work, I believe
these could either be turned back into static arrays or moved
to OpenJ9, but this commit is already pretty complicated so
I have left that cleanup for some other intrepid soul.

This change completely eliminates two of the more complex files
required in compiler extensions and simplifies how compiler
extensions (downstream projects) can choose the desired optimizer
configuration.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>