Restructure optimizer for JitBuilder and compilertest
We have two competing goals for the contents of the optimizer
component: 1) we want to test everything, and 2) it would be
nice not to have to include everything if not everything is
needed. The Optimizer class currently pulls a bit of a swizzle
by using a macro to rip out all the default optimization and
group initializations so that a compiler extension can define
its own optimizer. Unfortunately, this puts all the onus onto
the compiler extension to know what it wants if it doesn't want
everything. That's what the jitbuilder compiler extension does.

Rather than this all or "tell me everything you want" approach,
we could create a couple of options available for downstream
compilers. I'm going to call them FullOptimizer (i.e. enable
everything) and SmallOptimizer (which coresponds to what jitbuilder
currently uses). In fact, FullOptimizer and SmallOptimzier are
extensible classes.  By default, Optimizer.hpp will make
TR::Optimizer derive from TR::FullOptimizer. A compiler extension
(like jitbuilder) can now get the simpler SmallOptimizer by
defining the macro OMR_OPTIMIZER_SMALL in the DEFINES parameter
for building the compiler.  Defining this macro also causes
only OMRSMallOptimizer to be included in the build (otherwise
OMRFullOptimizer is included and OMRSmallOptimizer is not
included). By default, a compiler extension will
get everything (which is what was happening previously). I have
gotten rid of the original macro.

This change eliminates two of the more complex files required
in the jitbuilder compiler extension and simplifies how
compiler extensions can choose.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>