Prepare for OMR Optimizer consolidation with JitBuilder
JitBuilder currently overrides the set of optimizations in OMROptimizer
so that it can create smaller compilers by only including the optimizations
that give the biggest improvements.

Rather than this "override" approach, it would be better to split the
Optimizer implementation into a "minimal" set that JitBuilder can use
along with the more "full" set that other compiler extensions (like
OpenJ9) can use. This commit basically renames the current Optimizer
as MinOptimizer and creates a composable TR::MinOptimizer class. It
adjusts TR::Optimizer to be a subclass of TR::MinOptimizer. A
subsequent commit will reintroduce an OMROptimizer class that
subclasses TR::MinOptimizer so that we can shift optimization
initialization between these classes to make MinOptimizer smaller.

Some direct references to OMR::Optimizer were adjusted in this
commit to instead use the TR::Optimizer reference.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>