Create FullOptimizer extensible class subclassing TR::MinOptimizer
Create a FullOptimizer extensible class that inherits from
TR::MinOptimizer where we can move optimization initialization
that should not be in MinOptimizer. Also, update TR::Optimizer
to inherit from TR::FullOptimizer so default consumers will
get the FullOptimizer experience. In this commit, all of the
optimization initializations are still performed by
MinOptimizer, but we'll change that in a subsequent commit.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>