jb2: move some omrgen files around
Shift the more "private" OMR compiler related implementation files
directly into the omrgen/compiler directory and build them as part
of that directory. This includes OMRIlGen.cpp/hpp and
OMRCodeGenerator.cpp as well as OMRJB.cpp/hpp .

This has the advantage that none of the files in the omrgen/
directory itself include OMR compiler specific header files or
reference OMR compiler implementation (that's all confined to
the compiler directory).

I noticed that omrgen/Jit.cpp was not removed when it was
modified and then (supposedly) shifted into the
omrgen/compiler/control directory. Since it's the file in
that lower directory that's built now and not the one in
omrgen/, I removed the old file.

With this change, I also updated the default component
inclusions in jb2/CMakeLists.txt so that the omrgen tests can
be built and run. The hack to not include the Base component
tests is still present, and will remain until the jb extension
is fixed to build properly.

I also snuck in an indentation fix in jb/CMakeLists.txt that
has no functional impact to anything.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>