Prevent localSlot accessing array out of bounds
Because the slot can be for a local, it doesn't necessarily reside
within the parameter names array which is dereferenced here. The
easiest solution is to only access the parameter names array when
the slot is for a parameter. Without more knowledge about the
names of locals (maybe something that should be added later),
this fix returns a generic "Local #<slot>" string for the names
of locals as at least a valid and useful default answer.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>